full_data_shape in tools.hdmf
