+++ src/neuroconv/tools/hdmf.py
@@ -4,7 +4,12 @@ import math
 import warnings
 
 import numpy as np
+from hdmf.build.builders import (
+    BaseBuilder,
+    LinkBuilder,
+)
 from hdmf.data_utils import GenericDataChunkIterator as HDMFGenericDataChunkIterator
+from hdmf.utils import get_data_shape
 
 
 class GenericDataChunkIterator(HDMFGenericDataChunkIterator):  # noqa: D101
@@ -153,3 +158,151 @@ class SliceableDataChunkIterator(GenericDataChunkIterator):
 
     def _get_data(self, selection: tuple[slice]) -> np.ndarray:
         return self.data[selection]
+
+
+def get_full_data_shape(
+    dataset: GenericDataChunkIterator | np.ndarray | list,
+    location_in_file: str,
+    builder: BaseBuilder | None = None,
+):
+    """Get the full shape of the dataset at the given location in the file.
+
+    Parameters
+    ----------
+    dataset : hdmf.data_utils.GenericDataChunkIterator | np.ndarray | list
+        The dataset to get the shape of.
+    location_in_file : str
+        The location of the dataset within the NWBFile, e.g. 'acquisition/ElectricalSeries/data'.
+    builder : hdmf.build.builders.BaseBuilder | None
+        The builder object that would be used to construct the NWBFile object. If None, the dataset is assumed to NOT
+        have a compound dtype.
+
+    Notes
+    -----
+    This function is used instead of hdmf.utils.get_data_shape() to handle datasets with compound dtypes. Currently, if
+    a dataset has a compound dtype in NWB, the builder will write it as (len(dataset,), but hdmf.utils.get_data_shape()
+    will return the shape of the dataset as if it were a regular single-dtype array (ex. (N, M) instead of (N,)).
+    """
+    if builder is not None and has_compound_dtype(builder=builder, location_in_file=location_in_file):
+        return (len(dataset),)
+    return get_data_shape(data=dataset)
+
+
+def has_compound_dtype(builder: BaseBuilder, location_in_file: str) -> bool:
+    """
+    Determine if the dataset at the given location in the file has a compound dtype.
+
+    Parameters
+    ----------
+    builder : hdmf.build.builders.BaseBuilder
+        The builder object that would be used to construct the NWBFile object.
+    location_in_file : str
+        The location of the dataset within the NWBFile, e.g. 'acquisition/ElectricalSeries/data'.
+
+    Returns
+    -------
+    bool
+        Whether the dataset has a compound dtype.
+    """
+    dataset_builder = get_dataset_builder(builder, location_in_file)
+    return isinstance(dataset_builder.dtype, list)
+
+
+def get_dataset_builder(builder: BaseBuilder, location_in_file: str) -> BaseBuilder:
+    """Find the appropriate sub-builder for the dataset at the given location in the file.
+
+    This function will traverse the groups in the location_in_file until it reaches a DatasetBuilder,
+    and then return that builder.
+
+    Parameters
+    ----------
+    builder : hdmf.build.builders.BaseBuilder
+        The builder object that would be used to construct the NWBFile object.
+    location_in_file : str
+        The location of the dataset within the NWBFile, e.g. 'acquisition/ElectricalSeries/data'.
+
+    Returns
+    -------
+    hdmf.build.builders.BaseBuilder
+        The builder object for the dataset at the given location.
+
+    Raises
+    ------
+    ValueError
+        If the location_in_file is not found in the builder.
+
+    Notes
+    -----
+    Items in defined top-level places like electrodes may not be in the groups of the nwbfile-level builder,
+    but rather in hidden locations like general/extracellular_ephys/electrodes.
+    Also, some items in these top-level locations may interrupt the order of the location_in_file.
+    For example, when location_in_file is 'stimulus/AcousticWaveformSeries/data', the builder for that dataset is
+    located at 'stimulus/presentation/AcousticWaveformSeries/data'.
+    For this reason, we recursively search for the appropriate sub-builder for each name in the location_in_file.
+    Also, the first name in location_in_file is inherently suspect due to the way that the location is determined
+    in _find_location_in_memory_nwbfile(), and may not be present in the builder. For example, when location_in_file is
+    'lab_meta_data/fiber_photometry/fiber_photometry_table/location/data', the builder for that dataset is located at
+    'general/fiber_photometry/fiber_photometry_table/location/data'.
+    """
+    split_location = iter(location_in_file.split("/"))
+    name = next(split_location)
+
+    if _find_sub_builder(builder, name) is None:
+        name = next(split_location)
+
+    while name not in builder.datasets and name not in builder.links:
+        builder = _find_sub_builder(builder, name)
+        if builder is None:
+            raise ValueError(f"Could not find location '{location_in_file}' in builder ({name} is missing).")
+        try:
+            name = next(split_location)
+        except StopIteration:
+            raise ValueError(f"Could not find location '{location_in_file}' in builder ({name} is not a dataset).")
+    builder = builder[name]
+    if isinstance(builder, LinkBuilder):
+        builder = builder.builder
+    return builder
+
+
+def _find_sub_builder(builder: BaseBuilder, name: str) -> BaseBuilder:
+    """Search breadth-first for a sub-builder by name in a builder object.
+
+    Parameters
+    ----------
+    builder : hdmf.build.builders.BaseBuilder
+        The builder object to search for the sub-builder in.
+    name : str
+        The name of the sub-builder to search for.
+
+    Returns
+    -------
+    hdmf.build.builders.BaseBuilder
+        The sub-builder with the given name, or None if it could not be found.
+    """
+    sub_builders = list(builder.groups.values())
+    return _recursively_search_sub_builders(sub_builders=sub_builders, name=name)
+
+
+def _recursively_search_sub_builders(sub_builders: list[BaseBuilder], name: str) -> BaseBuilder:
+    """Recursively search for a sub-builder by name in a list of sub-builders.
+
+    Parameters
+    ----------
+    sub_builders : list[hdmf.build.builders.BaseBuilder]
+        The list of sub-builders to search for the sub-builder in.
+    name : str
+        The name of the sub-builder to search for.
+
+    Returns
+    -------
+    hdmf.build.builders.BaseBuilder
+        The sub-builder with the given name, or None if it could not be found.
+    """
+    sub_sub_builders = []
+    for sub_builder in sub_builders:
+        if sub_builder.name == name:
+            return sub_builder
+        sub_sub_builders.extend(list(sub_builder.groups.values()))
+    if len(sub_sub_builders) == 0:
+        return None
+    return _recursively_search_sub_builders(sub_builders=sub_sub_builders, name=name)
